pyproject.toml
[tool.poetry]
name = "passive-sound-localization"
version = "0.1.0"
description = ""
authors = ["John Finberg <johnbfinberg@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
numpy = "^2.1.1"
scipy = "^1.14.1"
scikit-learn = "^1.5.2"
matplotlib = "^3.9.2"
pyroomacoustics = "^0.7.7"
pyaudio = "^0.2.14"
wave = "^0.0.2"
hydra-core = "^1.3.2"
webrtcvad-wheels = "^2.0.14"
openai = "^1.51.0"
python-dotenv = "^1.0.1"
websockets = "^13.1"

[tool.poetry.group.test.dependencies]
pytest = "^8.3.3"
pytest-asyncio = "^0.24.0"

[tool.pytest.ini_options]
pythonpath = [
    "passive_sound_localization"
]

[tool.poetry.scripts]
docker = "scripts.run:docker"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"